src/requirements.txt
# Core dependencies
opencv-python>=4.8.0
pydantic>=2.0.0
pyyaml>=6.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Optional dependencies
# tqdm>=4.65.0  # Enhanced progress bars (optional)
